+++ src/test/resources/datacache/TomboloData/uk.gov.ons/9262535b-efe1-3ea4-b6aa-99e895e95fef.csv
@@ -1,0 +1,2 @@
+DATE_NAME,GEOGRAPHY_NAME,GEOGRAPHY_CODE,VARIABLE_NAME,MEASURES_NAME,OBS_VALUE,OBS_STATUS_NAME
+Oct 2016-Sep 2017,Barking and Dagenham,E09000002,Employment rate - aged 16-64,Variable,67,Normal Value
