+++ src/test/resources/datacache/TomboloData/uk.gov.ons/eff8daca-8ae6-30f5-aae1-00d2f8d21210.csv
@@ -1,0 +1,2 @@
+DATE_NAME,GEOGRAPHY_NAME,GEOGRAPHY_CODE,VARIABLE_NAME,MEASURES_NAME,OBS_VALUE,OBS_STATUS_NAME
+Oct 2016-Sep 2017,Barking and Dagenham,E09000002,Unemployment rate - aged 16-64 ,Variable,9.4,Normal Value
